Reorganise to put processor blocks in blocks, and recycle for tests
