+++ tests/test_summer_imsub_pipeline.py
@@ -1,64 +1,26 @@
 import unittest
 import logging
-from winterdrp.processors.autoastrometry import AutoAstrometry
-from winterdrp.processors.astromatic import Sextractor, Scamp, Swarp
-from winterdrp.pipelines.summer.summer_files import sextractor_photometry_config, psfex_config_path
-from winterdrp.processors.utils import ImageSaver
 from winterdrp.processors.utils.image_loader import ImageLoader
-from winterdrp.processors.utils.image_selector import ImageSelector, ImageBatcher
-from winterdrp.processors.photcal import PhotCalibrator
-from winterdrp.processors import MaskPixels, BiasCalibrator, FlatCalibrator
-from winterdrp.processors.csvlog import CSVLog
-from winterdrp.paths import core_fields, base_name_key
-from winterdrp.pipelines.summer.summer_pipeline import load_proc_summer_image, summer_pixel_scale, \
-    summer_reference_sextractor, summer_reference_image_generator, summer_reference_psfex, summer_reference_image_resampler, SummerPipeline
+from winterdrp.processors.utils.image_selector import ImageSelector
+from winterdrp.paths import base_name_key
+from winterdrp.pipelines.summer.summer_pipeline import SummerPipeline
 from winterdrp.downloader.get_test_data import get_test_data_dir
-from winterdrp.processors.reference import Reference
-from winterdrp.processors.astromatic import PSFex
-from winterdrp.processors.zogy.zogy import ZOGY, ZOGYPrepare, default_summer_catalog_purifier
+from winterdrp.pipelines.summer.blocks import subtract
+from winterdrp.pipelines.summer.load_summer_image import load_proc_summer_image
 
 logger = logging.getLogger(__name__)
 
 test_data_dir = get_test_data_dir()
 
 test_configuration = [
-            ImageLoader(
-                input_img_dir=test_data_dir,
-                input_sub_dir='processed',
-                load_image=load_proc_summer_image
-            ),
-            ImageBatcher(split_key=base_name_key),
-            ImageSelector(('OBSTYPE', 'SCIENCE')),
-            # ImageSelector(('FILTER', ['u'])),
-            ImageSelector((base_name_key, "SUMMER_20220816_042349_Camera0.resamp.fits")),
-            Reference(
-                ref_image_generator=summer_reference_image_generator,
-                ref_psfex=summer_reference_psfex,
-                ref_sextractor=summer_reference_sextractor,
-                ref_swarp_resampler=summer_reference_image_resampler
-            ),
-            Sextractor(
-                output_sub_dir='subtract',
-                cache=False,
-                write_regions_file=True,
-                **sextractor_photometry_config
-            ),
-            PSFex(config_path=psfex_config_path,
-                  output_sub_dir="subtract",
-                  norm_fits=True),
-            ZOGYPrepare(output_sub_dir="subtract", sci_zp_header_key='ZP_AUTO', catalog_purifier=default_summer_catalog_purifier),
-            ZOGY(output_sub_dir="subtract"),
-            # DetectCandidates(output_sub_dir="subtract",
-            #                  cand_det_sextractor_config='winterdrp/pipelines/summer/summer_imsub_files/config/photomCat.sex',
-            #                  cand_det_sextractor_nnw='winterdrp/pipelines/summer/summer_imsub_files/config/default.nnw',
-            #                  cand_det_sextractor_filter='winterdrp/pipelines/summer/summer_imsub_files/config/default.conv',
-            #                  cand_det_sextractor_params='winterdrp/pipelines/summer/summer_imsub_files/config/Scorr.param'),
-            # RegionsWriter(output_dir_name='candidates'),
-            # PSFPhotometry(),
-            # AperturePhotometry(aper_diameters=[8, 40], cutout_size_aper_phot=100, bkg_in_diameters=[25, 90],
-            #                    bkg_out_diameters=[40, 100], col_suffix_list=['', 'big']),
-            # DataframeWriter(output_dir_name='candidates'),
-        ]
+    ImageLoader(
+        input_img_dir=test_data_dir,
+        input_sub_dir='processed',
+        load_image=load_proc_summer_image
+    ),
+    ImageSelector((base_name_key, "SUMMER_20220816_042349_Camera0.resamp.fits")),
+] + subtract
+
 
 expected_values = {
     'SCORSTD': 1.120988782614284,
