+++ winterdrp/pipelines/summer/config/files/Scorr.param
@@ -1,0 +1,134 @@
+NUMBER
+
+#FLUX_ISO
+#FLUXERR_ISO
+#MAG_ISO
+#MAGERR_ISO
+
+#FLUX_ISOCOR
+#FLUXERR_ISOCOR
+#MAG_ISOCOR
+#MAGERR_ISOCOR
+
+FLUX_APER(1)
+FLUXERR_APER(1)
+#MAG_APER(1)
+#MAGERR_APER(1)
+
+FLUX_AUTO
+FLUXERR_AUTO
+#MAG_AUTO
+#MAGERR_AUTO
+
+#FLUX_BEST
+#FLUXERR_BEST
+#MAG_BEST
+#MAGERR_BEST
+
+#KRON_RADIUS
+FLUX_RADIUS
+#BACKGROUND
+
+#THRESHOLD
+#MU_THRESHOLD
+FLUX_MAX
+#MU_MAX
+#ISOAREA_IMAGE
+#ISOAREA_WORLD
+
+#XMIN_IMAGE
+#YMIN_IMAGE
+#XMAX_IMAGE
+#YMAX_IMAGE
+
+X_IMAGE
+Y_IMAGE
+XPEAK_IMAGE
+YPEAK_IMAGE
+#XWIN_IMAGE
+#YWIN_IMAGE
+ERRX2_IMAGE
+ERRY2_IMAGE
+#ERRAWIN_IMAGE
+#ERRBWIN_IMAGE
+#ERRTHETAWIN_IMAGE
+#X_WORLD
+#Y_WORLD
+#ALPHA_SKY
+#DELTA_SKY
+ALPHA_J2000
+DELTA_J2000
+#ALPHA_B1950
+#DELTA_B1950
+
+#X2_IMAGE
+#Y2_IMAGE
+#XY_IMAGE
+#X2_WORLD
+#Y2_WORLD
+#XY_WORLD
+
+#CXX_IMAGE
+#CYY_IMAGE
+#CXY_IMAGE
+#CXX_WORLD
+#CYY_WORLD
+#CXY_WORLD
+
+A_IMAGE
+B_IMAGE
+#A_WORLD
+#B_WORLD
+
+#THETA_IMAGE
+#THETA_WORLD
+#THETA_SKY
+#THETA_J2000
+#THETA_B1950
+
+ELONGATION
+#ELLIPTICITY
+
+#ERRX2_IMAGE
+#ERRY2_IMAGE
+#ERRXY_IMAGE
+#ERRX2_WORLD
+#ERRY2_WORLD
+#ERRXY_WORLD
+
+#ERRCXX_IMAGE
+#ERRCYY_IMAGE
+#ERRCXY_IMAGE
+#ERRCXX_WORLD
+#ERRCYY_WORLD
+#ERRCXY_WORLD
+
+#ERRA_IMAGE
+#ERRB_IMAGE
+#ERRA_WORLD
+#ERRB_WORLD
+
+#ERRTHETA_IMAGE
+#ERRTHETA_WORLD
+#ERRTHETA_SKY
+#ERRTHETA_J2000
+#ERRTHETA_B1950
+
+FWHM_IMAGE
+FWHM_WORLD
+
+#ISO0
+#ISO1
+#ISO2
+#ISO3
+#ISO4
+#ISO5
+#ISO6
+#ISO7
+
+FLAGS
+#IMAFLAGS_ISO
+#NIMAFLAGS_ISO(1)
+
+#CLASS_STAR
+#VIGNET(5,5)
