+++ winterdrp/pipelines/summer/config/files/photom.param
