+++ winterdrp/pipelines/summer/config/files/photomCat.sex
