+++ winterdrp/pipelines/summer/config/files/scamp.conf
