+++ winterdrp/pipelines/summer/config/schema/detfiles.sql
