+++ winterdrp/pipelines/summer/config/schema/dithers.sql
