+++ winterdrp/pipelines/summer/config/schema/exposures.sql
