+++ winterdrp/pipelines/summer/config/schema/nights.sql
